torrent: added Mockito

+++ src/main/java/ru/spbau/javacourse/torrent/utils/GlobalConstants.java
@@ -8,7 +8,7 @@ public class GlobalConstants {
     public final static String DOWNLOAD_DIR = "/downloads";
     public final static String CLIENT_DB_NAME = "CLIENT_FILE_DB";
     public final static String DEFAULT_HOST = "localhost";
-    public final static int TRACKER_PORT = 8081;
-    public final static int CLIENT_PORT_FST = 8841;
-    public final static int CLIENT_PORT_SND = 8842;
+    public final static short TRACKER_PORT = 8081;
+    public final static short CLIENT_PORT_FST = 8841;
+    public final static short CLIENT_PORT_SND = 8842;
 }
